# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления виртуальными кошельками с поддержкой конкурентных операций",
        "title": "Wallet Service API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/DisasterWoman/wallet-service"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/wallet": {
            "post": {
                "description": "Выполняет операцию пополнения или списания средств",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Изменить баланс кошелька",
                "parameters": [
                    {
                        "description": "Данные операции",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное выполнение операции",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Конфликт (недостаточно средств или кошелек не найден)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/{walletId}": {
            "get": {
                "description": "Возвращает текущий баланс указанного кошелька",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Получить баланс кошелька",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID кошелька",
                        "name": "walletId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Баланс кошелька",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Кошелек не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Возвращает статус работы сервиса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Проверка здоровья сервиса",
                "responses": {
                    "200": {
                        "description": "Сервис работает",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.OperationRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "operationType": {
                    "$ref": "#/definitions/models.OperationType"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.OperationType": {
            "type": "string",
            "enum": [
                "DEPOSIT",
                "WITHDRAW"
            ],
            "x-enum-varnames": [
                "Deposit",
                "Withdraw"
            ]
        }
    }
}